# swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/indego-data-fetch-and-store-it-db": {
            "post": {
                "description": "Fetches data from the Indego GeoJSON station status API and stores it in PostgreSQL.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indego - Weather"
                ],
                "summary": "Fetch and store Indego station data",
                "responses": {}
            }
        },
        "/api/v1/stations": {
            "get": {
                "description": "Retrieve the snapshot of all stations at a specified time, along with the weather data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indego - Weather"
                ],
                "summary": "Get station snapshot at a specific time",
                "parameters": [
                    {
                        "type": "string",
                        "description": "UTC timestamp (ISO 8601) for which to retrieve the station and weather snapshot.",
                        "name": "at",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
        "/api/v1/stations/{kioskId}": {
            "get": {
                "description": "Retrieve the snapshot of a specific station at a specified time, along with the weather data.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Indego - Weather"
                ],
                "summary": "Get station snapshot for a specific kiosk",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The kiosk ID of the station to retrieve.",
                        "name": "kioskId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "UTC timestamp (ISO 8601) for which to retrieve the station and weather snapshot.",
                        "name": "at",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {}
            }
        }
    }
}